requirements.txt
autopep8==1.4.4
certifi==2019.3.9
chardet==3.0.4
entrypoints==0.3
flake8==3.7.7
idna==2.8
mccabe==0.6.1
pkg-resources==0.0.0
pycodestyle==2.5.0
pyflakes==2.1.1
python-dotenv==0.10.3
requests==2.22.0
urllib3==1.25.3
